20-03-23

+++ part4/blogList/controllers/blogs.js
@@ -1,50 +1,61 @@
 require('express-async-errors')
 const blogsRouter = require('express').Router()
+const { userExtractor } = require('../utils/middleware')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  // Blog
-  //   .find({})
-  //   .then(blogs => {
-  //     response.json(blogs)
-  //   }).catch(next)
-
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {
+    username: 1
+  })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  // const blog = new Blog(request.body)
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const newBlog = request.body
+  const user = request.user
 
   if (newBlog.title === undefined || newBlog.url === undefined) {
     return response.status(400).json('no title/url')
   }
 
+  const userPosting = await User.findById(user.id)
+
   const blog = new Blog({
     title: newBlog.title,
     author: newBlog.author,
     url: newBlog.url,
-    likes: newBlog.likes || 0
+    likes: newBlog.likes || 0,
+    user: userPosting._id
   })
 
-  // blog
-  //   .save()
-  //   .then(result => {
-  //     response.status(201).json(result)
-  //   })
-  //   .catch(next)
+  const savedBlog = await blog.save()
+  userPosting.blogs = userPosting.blogs.concat(savedBlog._id)
+  await userPosting.save()
 
-  const result = await blog.save()
-  response.status(201).json(result)
+  response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (user.id !== blog.user.toString()) {
+    return response.status(400).json('blog can be deleted only by creator')
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (user.id !== blog.user.toString()) {
+    return response.status(400).json('blog can be deleted only by creator')
+  }
+
   const updatedBlog = {
     likes: request.body.likes
   }
